1050fe9cae88643ebe18436bc99a02f2c1c9961a Remove unnecessary borrow (Tobin C. Harding)
39667093366dc70692f3d7c147ca9619b395fc4a Use is_none() (Tobin C. Harding)
d1920525192c8e6011073ca2ab0e37ac974b6edf Remove unnecessary dereference (Tobin C. Harding)
624cda07b3f1d36928f92b904368c3ad4de16ec4 Remove unnecessary casts (Tobin C. Harding)

Pull request description:

  Clippy has been updated and new warnings are being triggered in our codebase. This PR does all warnings using nightly since they all looked like reasonable things to fix.

  Needed for CI to pass in other open PRs.

ACKs for top commit:
  Kixunil:
    ACK 1050fe9cae88643ebe18436bc99a02f2c1c9961a
  sanket1729:
    ACK 1050fe9cae88643ebe18436bc99a02f2c1c9961a.

Tree-SHA512: 7dcfb6a72a0aae51b49b417bb94cbe1becb1095d1bf0011921b1834a10f792cfcdeee37993ab9b103bd2dfcc9cd3c26cd7f1bb80b06b0d1aa4aaa454bfb0b3f0

+++ bitcoin/src/bip152.rs
@@ -284,7 +284,7 @@ impl Decodable for BlockTransactionsRequest {
                 // Since the number of indices ultimately represent transactions,
                 // we can limit the number of indices to the maximum number of
                 // transactions that would be allowed in a vector.
-                let byte_size = (nb_indexes as usize)
+                let byte_size = (nb_indexes)
                     .checked_mul(mem::size_of::<Transaction>())
                     .ok_or(encode::Error::ParseFailed("Invalid length"))?;
                 if byte_size > encode::MAX_VEC_SIZE {
